refactor: move sort_by_user_query to result_utils.py
